pyproject.toml
[project]
name = "roverteam-spot"
version = "0.1.0"
description = ""
authors = [
    {name = "Aston Bolwell"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "json5 (>=0.13.0,<0.14.0)",
    "pyzmq (>=25.0.0,<27.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "aiortc (>=1.14.0,<2.0.0)",
    "av (>=16.1.0,<17.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "opencv-python (>=4.13.0.90,<5.0.0.0); sys_platform == 'win32'"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false